Milestone Alpha
I added a door, furniture, and other interactions (the door unlocks when the puzzle is solved, and exiting through the door ends the game). I also added post-it notes, which will be used in future updates. When clicked, the post-its zoom in to reveal clues necessary for solving other puzzles.